+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version History
 -   Lights can be now part of `OSPGroup` and thus instanced like geometries
     and volumes
 -   Added support for lights Motion Blur (with the path tracer)
+-   Support rolling shutter effect
 -   Fix illumination from emissive quad meshes
 
 ### Changes in v2.7.1: